# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Seu Zuca — Marketplace B2B de Materiais de Construção
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /supplier/orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [supplier]
      summary: Atualizar status do pedido
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: Listar pedidos do comprador
      responses:
        "200":
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Criar pedido (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderWithCheckout"
        "400":
          description: Erro na criação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Obter pedido por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # QUOTES
...
  /cep/{cep}:
    get:
      operationId: lookupCep
      tags: [cep]
      summary: Consultar CEP via ViaCEP
      parameters:
        - name: cep
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dados do CEP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CepData"

  # ADMIN
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - message
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantidade:
          type: integer
        precoUnitario:
          type: number
        subtotal:
          type: number
      required:
        - id
        - productId
        - productName
        - quantidade
        - precoUnitario
        - subtotal
    Order:
      type: object
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        status:
          type: string
          enum: [pendente, em_separacao, enviado, entregue, cancelado]
        total:
          type: number
        comissao:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        addressId:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - buyerId
        - status
        - total
        - createdAt
    OrderWithCheckout:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            checkoutUrl:
              type: string
            stripeSessionId:
              type: string
    CreateOrderBody:
      type: object
      properties:
        addressId:
          type: integer
        observacoes:
          type: string
      required:
        - addressId
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pendente, em_separacao, enviado, entregue, cancelado]
      required:
        - status

    # QUOTES
...
    CepData:
      type: object
      properties:
        cep:
          type: string
        logradouro:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
        ibge:
          type: string
      required:
        - cep

    # STATS
...