# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Seu Zuca — Marketplace B2B de Materiais de Construção
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  buyerId: number;
  buyerName?: string;
  supplierId?: number;
  supplierName?: string;
  status: OrderStatus;
  total: number;
  comissao?: number;
  items?: OrderItem[];
  addressId?: number;
  createdAt: string;
  updatedAt?: string;
}